US1699041: create class to contain logic for getting/setting attribute values
