+++ access-checkout/src/main/java/com/worldpay/access/checkout/ui/AccessCheckoutEditText.kt
@@ -16,7 +16,6 @@ import androidx.annotation.ColorInt
 import androidx.annotation.RequiresApi
 import androidx.annotation.StyleRes
 import androidx.core.widget.TextViewCompat
-import com.worldpay.access.checkout.R
 import kotlin.random.Random
 
 class AccessCheckoutEditText internal constructor(
@@ -33,7 +32,9 @@ class AccessCheckoutEditText internal constructor(
         orientation = VERTICAL
         this.editText.id = this.id + editTextPartialId
 
-        val attributes: TypedArray? = attrs?.let { setAttributes(context, attrs, defStyle) }
+        val attributeValues = AttributeValues(context, attrs, this.editText)
+
+        val attributes: TypedArray? = attrs?.let { attributeValues.setAttributes() }
 
         addView(this.editText)
 
@@ -183,148 +184,4 @@ class AccessCheckoutEditText internal constructor(
         super.onRestoreInstanceState(bundledState.getBundle("superState"))
         editText.onRestoreInstanceState(bundledState.getBundle("editTextState"))
     }
-
-    private fun setAttributes(
-        context: Context,
-        attrs: AttributeSet?,
-        defStyle: Int
-    ): TypedArray {
-        val styleAttributes: TypedArray =
-            context.obtainStyledAttributes(attrs, R.styleable.AccessCheckoutEditText, defStyle, 0)
-
-        val textColor = getTextColorAttribute(styleAttributes)
-        setTextColourAttribute(textColor)
-
-        val hint = getHintAttribute(styleAttributes)
-        setHintAttribute(hint)
-
-        val ems = getEmsAttribute(styleAttributes)
-        setEmsAttribute(ems)
-
-        val hintTextColour = getHintTextColourAttribute(styleAttributes)
-        setHintTextColourAttribute(hintTextColour)
-
-        val imeOptions = getImeOptionsAttribute(styleAttributes)
-        setImeOptionsAttribute(imeOptions)
-
-        val cursorVisible = getCursorVisibleAttribute(styleAttributes)
-        setCursorVisibleAttribute(cursorVisible)
-
-        val textScaleX = getTextScaleXAttribute(styleAttributes)
-        setTextScaleXAttribute(textScaleX)
-
-        val textSize = getTextSizeAttribute(styleAttributes)
-        setTextSizeAttribute(textSize)
-
-        // set padding
-        val padding = getPaddingAttribute(styleAttributes).toInt()
-        val paddingLeft = getPaddingLeftAttribute(styleAttributes).toInt()
-        val paddingTop = getPaddingTopAttribute(styleAttributes).toInt()
-        val paddingRight = getPaddingRightAttribute(styleAttributes).toInt()
-        val paddingBottom = getPaddingBottomAttribute(styleAttributes).toInt()
-
-        setPaddingAttribute(paddingLeft, paddingTop, paddingRight, paddingBottom, padding)
-
-        return styleAttributes
-    }
-
-    private fun setPaddingAttribute(
-        paddingLeft: Int,
-        paddingTop: Int,
-        paddingRight: Int,
-        paddingBottom: Int,
-        padding: Int
-    ) {
-        var paddingLeft1 = paddingLeft
-        var paddingTop1 = paddingTop
-        var paddingRight1 = paddingRight
-        var paddingBottom1 = paddingBottom
-
-        paddingLeft1.takeIf { it == 0 }?.let { paddingLeft1 = editText.paddingLeft }
-        paddingTop1.takeIf { it == 0 }?.let { paddingTop1 = editText.paddingTop }
-        paddingRight1.takeIf { it == 0 }?.let { paddingRight1 = editText.paddingRight }
-        paddingBottom1.takeIf { it == 0 }?.let { paddingBottom1 = editText.paddingBottom }
-
-        if (padding != 0) {
-            paddingLeft1 = padding
-            paddingTop1 = padding
-            paddingRight1 = padding
-            paddingBottom1 = padding
-        }
-
-        editText.setPadding(paddingLeft1, paddingTop1, paddingRight1, paddingBottom1)
-    }
-
-    private fun setCursorVisibleAttribute(cursorVisible: Boolean) {
-        this.isCursorVisible = cursorVisible
-    }
-
-    private fun setHintAttribute(hint: String?) {
-        hint?.let {
-            this.setHint(it as CharSequence)
-        }
-    }
-
-    private fun setTextScaleXAttribute(textScaleX: Float) {
-        textScaleX.takeIf { it != 0F }?.let { this.textScaleX = textScaleX }
-    }
-
-    private fun setTextSizeAttribute(textSize: Float) {
-        textSize.takeIf { it != 0F }?.let { this.textSize = textSize }
-    }
-
-    private fun setImeOptionsAttribute(imeOptions: Int) {
-        imeOptions.takeIf { it != 0 }?.let { this.imeOptions = imeOptions }
-    }
-
-    private fun setHintTextColourAttribute(hintTextColour: Int) {
-        hintTextColour.takeIf { it != 0 }?.let { this.setHintTextColor(hintTextColour) }
-    }
-
-    private fun setEmsAttribute(ems: Int) {
-        ems.takeIf { it != 0 }?.let { this.setEms(ems) }
-    }
-
-    private fun setTextColourAttribute(textColor: Int) {
-        textColor.takeIf { it != 0 }?.let { this.setTextColor(textColor) }
-    }
-
-    private fun getPaddingBottomAttribute(styleAttributes: TypedArray) =
-        styleAttributes.getDimension(R.styleable.AccessCheckoutEditText_android_paddingBottom, 0.0F)
-
-    private fun getPaddingRightAttribute(styleAttributes: TypedArray) =
-        styleAttributes.getDimension(R.styleable.AccessCheckoutEditText_android_paddingRight, 0.0F)
-
-    private fun getPaddingTopAttribute(styleAttributes: TypedArray) =
-        styleAttributes.getDimension(R.styleable.AccessCheckoutEditText_android_paddingTop, 0.0F)
-
-    private fun getPaddingLeftAttribute(styleAttributes: TypedArray) =
-        styleAttributes.getDimension(R.styleable.AccessCheckoutEditText_android_paddingLeft, 0.0F)
-
-    private fun getPaddingAttribute(styleAttributes: TypedArray) =
-        styleAttributes.getDimension(R.styleable.AccessCheckoutEditText_android_padding, 0.0F)
-
-    private fun getHintAttribute(styleAttributes: TypedArray) =
-        styleAttributes.getString(R.styleable.AccessCheckoutEditText_android_hint)
-
-    private fun getTextSizeAttribute(styleAttributes: TypedArray) =
-        styleAttributes.getDimension(R.styleable.AccessCheckoutEditText_android_textSize, 0F)
-
-    private fun getTextScaleXAttribute(styleAttributes: TypedArray) =
-        styleAttributes.getFloat(R.styleable.AccessCheckoutEditText_android_textScaleX, 0F)
-
-    private fun getCursorVisibleAttribute(styleAttributes: TypedArray) =
-        styleAttributes.getBoolean(R.styleable.AccessCheckoutEditText_android_cursorVisible, true)
-
-    private fun getImeOptionsAttribute(styleAttributes: TypedArray) =
-        styleAttributes.getInt(R.styleable.AccessCheckoutEditText_android_imeOptions, 0)
-
-    private fun getHintTextColourAttribute(styleAttributes: TypedArray) =
-        styleAttributes.getColor(R.styleable.AccessCheckoutEditText_android_textColorHint, 0)
-
-    private fun getEmsAttribute(styleAttributes: TypedArray) =
-        styleAttributes.getInt(R.styleable.AccessCheckoutEditText_android_ems, 0)
-
-    private fun getTextColorAttribute(styleAttributes: TypedArray) =
-        styleAttributes.getColor(R.styleable.AccessCheckoutEditText_android_textColor, 0)
 }
